refactor: move out row in prep for dialog
